Check called on type instead of declaring type too
I noticed that disallowing `DateTime::RSS` doesn't work.

This is because `DateTime` extends `DateTimeInterface` and the `RSS` constant is defined on that interface.

`ClassConstantUsages` resolves the constant fetches to the declaring classes. That means that it would check if `DateTimeInterface::RSS` is allowed or not. That would never match.

To me this is a hidden concern that is not really what you expect as an end user.

To solve the problem, `ClassConstantUsages` now first checks the declaring constant (old behavior). If that doesn't give any errors, it checks the called on type for errors.