Move the flag config type check to paramFactory()
There's already a similar check there and PHPUnit 11 throws the following error when the check is done when the code is analyzed. That's because PHPStan sets its own error handler in \PHPStan\Analyser\FileAnalyser::analyseFile() ($this->collectErrors($analysedFiles)) but when there's an error like this, it doesn't reset it.

The PHPUnit error:
1) Spaze\PHPStan\Rules\Disallowed\Calls\FunctionCallsTypeStringParamsInvalidFlagsConfigTest::testException
Test code or tested code did not remove its own error handlers