Reproducer for #323
When I try it in user land you can see the problem.

```command
composer req phpstan/phpstan:2.1.11 && vendor/bin/phpstan --configuration=reproducer.neon

  15     Class Reproducer\EventRecorder is forbidden.
         🪪  disallowed.class
  25     Class Reproducer\EventRecorder is forbidden.
         🪪  disallowed.class
```

This is expected.

When you try it with:

```command
composer req phpstan/phpstan:2.1.12 && vendor/bin/phpstan --configuration=reproducer.neon

  15     Class Reproducer\EventRecorder is forbidden.
         🪪  disallowed.class
  25     Class Reproducer\EventRecorder is forbidden.
         🪪  disallowed.class
  45     Class Reproducer\EventRecorder is forbidden.
         🪪  disallowed.class
```

I cannot reproduce this problem in a unit test, because the test only tests a single Rule at a time.
Plus it might be related to the way the helper rules work and are mocked/faked in test.