Add shipmonk/dead-code-detector to detect you know what
Running two configurations because the detector may or may not detect different things when analyzing tests as well.

For example the services not registered in the DI container as reported in #330 wouldn't be detected when tests would be analyzed as well. So the "no tests" config is more important I think, and because it can't analyze the tests path, it needs to be an extra config file than the main phpstan.neon. But why not analyze the tests as well, right. I'm also using extra config files because neither of the configs need to run on all supported PHP versions.

Another possibility could be to use the usageExcluders config option, maybe
https://github.com/shipmonk-rnd/dead-code-detector#excluding-usages-in-tests